refactor!: reorganize Flux functions and rename files
