+++ Flux.Net/Flux.Net/FluxQueryBuilder.cs
@@ -1,4 +1,5 @@
-﻿using NodaTime;
+﻿using Flux.Net.Extensions;
+using NodaTime;
 using System;
 using System.Collections.Generic;
 using System.Text;
