+++ Flux.Net/Flux.Net/TimeUnit.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using Flux.Net.Extensions;
+using System.Collections.Generic;
 
 namespace Flux.Net
 {
